Adding correlation ids to accurately track proofs (#25)
* Adding correlation ids to accurately track proofs

Details:
Added clog_writert class to add correlation identifiers to all proof steps in Padstone.  This allows construction of a task tree showing the status of all tasks in real-time and also to more accurately track failures, especially when multiple proofs are in flight.  Finally, the identifiers are independent of the current lambda structure so are robust if we add/remove lambdas.

Adds three log entries per task: SUBMITTED, STARTED, and COMPLETED.
See clog_writert.py for more complete information

To report on proofs, I have added capability in proof.py to track proofs based on correlation identifiers, and also made the existing support in proof.py more robust if the correlation identifiers are absent (as with past proofs).

As a process thing, I had a previous CR out but the sequence of commits gets messy when doing snapshot-updates to test new functionality, so I created a separate branch with a cleaner commit history.  In future PRs, I will separate the branch for testing from the PR branch so it does not get messy.

Testing:
I have tested the approach against all of the failure modes
I could think of:
  - well-formedness errors for the python code for HandleWebhookLambda, InvokeBatchLambda, and prepare_source
  - missing entrypoints for the python code for HandleWebhookLambda, InvokeBatchLambda, and prepare_source
  - container errors for aws batch and codebuild
  - proof errors

One change involves taking the inlined handle_webhook code out of the github.yaml and adding it to a handle_webhook.py file in the standard lambda.zip file.  This change does not appreciatively add to latency (it is a wash) and opens the door to the removal of the InvokeBatchLambda lambda.  By doing so, we reduce latency by about 1 second.

Here are the results of a couple of runs with each approach:
Time from (old) multiple lambda w/inline Python (starting from canary): 6185 ms
Webhook execution time: 2811.51 ms

Time from (new) Webhook lambda with 17MB .zip file (starting from canary): 6180 ms
Webhook execution time: 2530 ms. <-- this is slightly surprising.

Time from (new) single 17MB lambda containing both HandleWebhook and InvokeBatchStatus (starting from canary): 5310 ms
Webhook execution time: 1603.34 ms

* Added timings and made proof.py more robust to missing information.

There is a chance that an AWS Batch job will not have a container if (for example) the docker image cannot be found.  This change handles this case.

Also, added support for timestamps for time tracking.

* Addressing PR comments.