Don't alarm on missing commits
This addresses a common reason for spurious alarms. GitHub users often
push a branch, and then force-push over that branch before the CodeBuild
stage had time to check out the first push. When this happens, GitHub
deletes the commits in the original branch. This means that attempting
to check out that commit, which is passed to CodeBuild via the GitHub
webhook, fails and causes an alarm.

This commit returns early from CodeBuild if the commit indicated in the
webhook does not exist in the repository.