Don't send VerificationErrorAlarm for missing hash (#37)
This commit prevents prepare_source.py (which runs during the CodeBuild
stage) from emitting a VerificationErrorAlarm when an exception is
raised. Emitting an alarm is unnecessary because CodeBuild itself will
emit a CodeBuildError when it catches the exception. In addition, the
error is misleading because no verification was attempted (a
VerificationError should only be emitted if there was a problem with a
proof run, while this error occurs before any proof runs are attempted).

One trigger for this problem is a push to a pull request, followed by a
force-push in quick succession. GitHub deletes the commit from the first
pull request, meaning that the attempt to check out that commit fails.
This raises an exception, which is handled by updating the GitHub
status, but the call to update_status incorrectly set no_metric_status
to False.